Relocated and refactored scripts pulled from the html; Created a components folder and moved  the load table of contents function into it.
